patch 7.4.839 Problem:    Compiler warning on 64-bit system. Solution:   Add cast to int. (Mike Williams)

+++ src/search.c
@@ -1801,7 +1801,7 @@ find_rawstring_end(linep, startpos, endpos)
     for (p = linep + startpos->col + 1; *p && *p != '('; ++p)
 	;
     delim_len = (p - linep) - startpos->col - 1;
-    delim_copy = vim_strnsave(linep + startpos->col + 1, delim_len);
+    delim_copy = vim_strnsave(linep + startpos->col + 1, (int)delim_len);
     if (delim_copy == NULL)
 	return FALSE;
     for (lnum = startpos->lnum; lnum <= endpos->lnum; ++lnum)
